requirements.txt
numpy
pandas
scikit-learn
opencv-python
scikit-image
torch==2.3.0
torchvision==0.18.0
mlflow
streamlit
pillow
matplotlib
requests
jupyter
